implementação do unit of work, otimização de algumas consultas e pop de confirmação

+++ client/src/app/messages/messages.component.ts
@@ -4,6 +4,7 @@ import { Pagination } from '../_models/pagination';
 import { MessageService } from '../_services/message.service';
 import { strings as stringsPtBr } from "ngx-timeago/language-strings/pt-br"
 import { TimeagoIntl } from 'ngx-timeago';
+import { ConfirmService } from '../_services/confirm.service';
 
 @Component({
   selector: 'app-messages',
@@ -18,7 +19,8 @@ export class MessagesComponent implements OnInit {
   pageSize = 5;
   loading = false;
 
-  constructor(private messageService: MessageService,  public intl: TimeagoIntl) {
+  constructor(private messageService: MessageService,  public intl: TimeagoIntl, 
+      private confirmService: ConfirmService) {
     intl.strings = stringsPtBr;
     intl.changes.next();
    }
@@ -37,11 +39,15 @@ export class MessagesComponent implements OnInit {
   }
 
   deleteMessage(id: number){
-    this.messageService.deleteMessage(id).subscribe(() =>{
-      this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
-    })
-  }
-
+    this.confirmService.confirm('Confirmar exclusão', 'Não pode ser desfeito').subscribe( result => {
+      if(result) {      
+        this.messageService.deleteMessage(id).subscribe(() => {
+        this.messages.splice(this.messages.findIndex(m => m.id === id), 1)
+      })
+    }
+  })
+}
+    
   pageChanged(event: any) {
     this.pageNumber = event.page;
     this.loadMessages();
